refactor(back): rework maps.service structure and status permissions logic
Nick: This commit is horrible, I'm really sorry. I fucked up and did this big move, then refactored status change stuff. Recommitting cleaning would take hours.

If you try to review individually you'll see a bunch of stuff you just reviewed being rewritten - hence why I'm gonna suggest you review this branch by "Files Changed".
