+++ apps/backend/src/app/modules/maps/map-credits.service.ts
@@ -2,6 +2,7 @@ import {
   BadRequestException,
   ConflictException,
   ForbiddenException,
+  forwardRef,
   Inject,
   Injectable,
   NotFoundException
@@ -30,6 +31,7 @@ import { ConfigService } from '@nestjs/config';
 export class MapCreditsService {
   constructor(
     @Inject(EXTENDED_PRISMA_SERVICE) private readonly db: ExtendedPrismaService,
+    @Inject(forwardRef(() => MapsService))
     private readonly mapsService: MapsService,
     private readonly config: ConfigService
   ) {}
